Nix mysleep().
Just use sleep() where only whole-second resolution is used
to delay between reconnect-attempts; and nanosleep where
sub-second resolution is used.

sleep() should actually be OK, because we don't intermix those calls
with alarm() or any other signals or itimer functions.

nanosleep() should be OK because POSIX.1-2001 requires
that it not have the crazy signal interactions
that its predecessors are known for.